+++ packages/imperative/src/rest/src/client/AbstractRestClient.ts
@@ -25,7 +25,7 @@ import { HTTP_VERB } from "./types/HTTPVerb";
 import { ImperativeExpect } from "../../../expect/src/ImperativeExpect";
 import { Session } from "../session/Session";
 import * as path from "path";
-import { IRestClientError } from "./doc/IRestClientError";
+import { completionTimeoutErrorMessage, IRestClientError } from "./doc/IRestClientError";
 import { RestClientError } from "./RestClientError";
 import { Readable, Writable } from "stream";
 import { IO } from "../../../io";
@@ -341,10 +341,18 @@ export abstract class AbstractRestClient {
                 }
             }
 
+            // Set up the request timeout
+            if (this.mSession.ISession.requestCompletionTimeout && this.mSession.ISession.requestCompletionTimeout > 0) {
+                clientRequest.setTimeout(this.mSession.ISession.requestCompletionTimeout);
+            }
+
             clientRequest.on("timeout", () => {
                 if (clientRequest.socket.connecting) {
                     // We timed out. Destroy the request.
                     clientRequest.destroy(new Error("Connection timed out. Check the host, port, and firewall rules."));
+                } else if (this.mSession.ISession.requestCompletionTimeout && this.mSession.ISession.requestCompletionTimeout > 0) {
+                    this.mSession.ISession.requestCompletionTimeoutCallback?.();
+                    clientRequest.destroy(new ImperativeError({msg: completionTimeoutErrorMessage}));
                 }
             });
 
@@ -359,6 +367,12 @@ export abstract class AbstractRestClient {
                     }).catch((err) => {
                         reject(err);
                     });
+                } else if (errorResponse instanceof ImperativeError && errorResponse.message === completionTimeoutErrorMessage) {
+                    reject(this.populateError({
+                        msg: "HTTP request timed out after connecting.",
+                        causeErrors: errorResponse,
+                        source: "timeout"
+                    }));
                 } else {
                     reject(this.populateError({
                         msg: "Failed to send an HTTP request.",
@@ -476,12 +490,25 @@ export abstract class AbstractRestClient {
 
         this.validateRestHostname(this.session.ISession.hostname);
 
-        let socketConnectTimeout: number;
         if (ImperativeConfig.instance.envVariablePrefix) {
-            const envValue = EnvironmentalVariableSettings.read(ImperativeConfig.instance.envVariablePrefix).socketConnectTimeout.value;
-            socketConnectTimeout = isNaN(Number(envValue)) ? undefined : Number(envValue);
+            const envValues = EnvironmentalVariableSettings.read(ImperativeConfig.instance.envVariablePrefix);
+            const socketConnectTimeout = envValues.socketConnectTimeout?.value;
+            const requestCompletionTimeout = envValues.requestCompletionTimeout?.value;
+
+            this.session.ISession.socketConnectTimeout ??= isNaN(Number(socketConnectTimeout)) ? undefined : Number(socketConnectTimeout);
+            if (this.session.ISession.socketConnectTimeout != null) {
+                Logger.getImperativeLogger().info(
+                    "Setting socket connection timeout ms: " + String(this.mSession.ISession.requestCompletionTimeout)
+                );
+            }
+
+            this.session.ISession.requestCompletionTimeout ??= isNaN(Number(requestCompletionTimeout)) ? undefined : Number(requestCompletionTimeout);
+            if (this.session.ISession.requestCompletionTimeout != null) {
+                Logger.getImperativeLogger().info(
+                    "Setting request completion timeout ms: " + String(this.mSession.ISession.requestCompletionTimeout)
+                );
+            }
         }
-        socketConnectTimeout = this.session.ISession.socketConnectTimeout ?? socketConnectTimeout;
 
         /**
          * HTTPS REST request options
@@ -502,7 +529,7 @@ export abstract class AbstractRestClient {
             port: this.session.ISession.port,
             rejectUnauthorized: this.session.ISession.rejectUnauthorized,
             // Timeout after failing to connect for 60 seconds, or sooner if specified
-            timeout: socketConnectTimeout
+            timeout: this.session.ISession.socketConnectTimeout
         };
 
         // NOTE(Kelosky): This cannot be set for http requests
@@ -935,6 +962,8 @@ export abstract class AbstractRestClient {
             detailMessage =
                 `HTTP(S) client encountered an error. Request could not be initiated to host.\n` +
                 `Review connection details (host, port) and ensure correctness.`;
+        } else if (finalError.source === "timeout") {
+            detailMessage = `HTTP(S) client encountered an error. Request timed out.`;
         } else {
             detailMessage =
                 `Received HTTP(S) error ${finalError.httpStatus} = ${http.STATUS_CODES[finalError.httpStatus]}.`;